+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ffi` feature to enable `repr(C)` on some structs and enums.
 - Implement MemoryAccess trait for `[u8]`, `&[u8]`, `[u16]` and `&[u16]`.
 - Line A and line F emulator exceptions.
+- Added `with_sr` function for `Registers`.
 
 ### Changed
 - Change license from LGPL-3.0 to MPL-2.0.